Refactor migrator classes into own modules and fix YAML file loading
