+++ nmdc_schema/migrators/README.md
@@ -1,0 +1,7 @@
+# Migrators
+
+## Overview
+
+This Python [package](https://docs.python.org/3/tutorial/modules.html#packages) (i.e. folder) contains
+[modules](https://docs.python.org/3/tutorial/modules.html#modules) (i.e. `.py` files) which contain
+[classes](https://docs.python.org/3/tutorial/classes.html) related to the migration of data between schema versions.
