+++ nmdc_schema/migrators/migrator_from_7_7_2_to_7_8_0.py
@@ -1,0 +1,33 @@
+import re
+import logging
+from .migrator_base import MigratorBase
+
+logger = logging.getLogger(__name__)
+
+# TODO: Capitalize the names of this constant, per: https://peps.python.org/pep-0008/#constants
+doi_url_pattern = r"^https?:\/\/[a-zA-Z\.]+\/10\."
+
+
+class Migrator_from_7_7_2_to_7_8_0(MigratorBase):
+    """Migrates data from schema 7.7.2 to 7.8.0"""
+
+    def __init__(self) -> None:
+        """Invokes parent constructor and populates collection-to-transformations map."""
+
+        super().__init__()
+
+        # Populate the "collection-to-transformers" map for this specific migration.
+        self.agenda = dict(
+            study_set=[self.replace_doi_field_with_award_dois_list_field],
+        )
+
+    def replace_doi_field_with_award_dois_list_field(self, study: dict) -> dict:
+        logger.info(f"Starting migration of {study['id']}")
+        if "doi" in study:
+            match = re.search(doi_url_pattern, study["doi"]["has_raw_value"])
+            if match:
+                start_index = match.end()
+                as_curie = f"doi:10.{study['doi']['has_raw_value'][start_index:]}"
+                study["award_dois"] = [as_curie]
+            del study["doi"]
+        return study
